Fixed list page to map page overlay issue
Due to the address of the map page being changed from lonodn_parkive.html to index.html the park overlay was not launching from the list page. This has now been fixed. Along with the landing popup no longer appearing when the map page is launched from a park name being clicked on in the list view.